Split subscribe pipeline into separate files.
