+++ src/subscribe/OrderMessages.js
@@ -1,0 +1,77 @@
+import { Utils } from 'streamr-client-protocol'
+
+import { pipeline } from '../utils/iterators'
+import PushQueue from '../utils/PushQueue'
+
+import resendStream from './resendStream'
+import { validateOptions } from './api'
+
+const { OrderingUtil } = Utils
+
+export default function OrderMessages(client, options = {}) {
+    const { gapFillTimeout, retryResendAfter } = client.options
+    const { streamId, streamPartition } = validateOptions(options)
+
+    const outStream = new PushQueue()
+
+    let done = false
+    const resendStreams = new Set()
+    const orderingUtil = new OrderingUtil(streamId, streamPartition, (orderedMessage) => {
+        if (!outStream.isWritable() || done) {
+            return
+        }
+
+        if (orderedMessage.isByeMessage()) {
+            outStream.end(orderedMessage)
+        } else {
+            outStream.push(orderedMessage)
+        }
+    }, async (from, to, publisherId, msgChainId) => {
+        if (done) { return }
+        // eslint-disable-next-line no-use-before-define
+        const resendMessageStream = await resendStream(client, {
+            streamId, streamPartition, from, to, publisherId, msgChainId,
+        })
+
+        try {
+            if (done) { return }
+            resendStreams.add(resendMessageStream)
+            await resendMessageStream.subscribe()
+            if (done) { return }
+
+            for await (const { streamMessage } of resendMessageStream) {
+                if (done) { return }
+                orderingUtil.add(streamMessage)
+            }
+        } finally {
+            resendStreams.delete(resendMessageStream)
+            await resendMessageStream.cancel()
+        }
+    }, gapFillTimeout, retryResendAfter)
+
+    const markMessageExplicitly = orderingUtil.markMessageExplicitly.bind(orderingUtil)
+
+    return Object.assign(pipeline([
+        // eslint-disable-next-line require-yield
+        async function* WriteToOrderingUtil(src) {
+            for await (const msg of src) {
+                orderingUtil.add(msg)
+            }
+        },
+        outStream,
+        async function* WriteToOrderingUtil(src) {
+            for await (const msg of src) {
+                yield msg
+            }
+        },
+    ], async (err) => {
+        done = true
+        orderingUtil.clearGaps()
+        resendStreams.forEach((s) => s.cancel())
+        resendStreams.clear()
+        await outStream.cancel(err)
+        orderingUtil.clearGaps()
+    }), {
+        markMessageExplicitly,
+    })
+}
